requirements.txt
python-dotenv
openai
youtube-transcript-api
pytube
httpx